requirements.txt
bs4==0.0.1
click==7.1.2
jwt==1.2.0
requests==2.25.1
python-dotenv==0.15.0